+++ castle/cms/indexing/hps.py
@@ -12,8 +12,6 @@ from wildcard.hps.interfaces import IQueryAssembler
 from wildcard.hps.opensearch import WildcardHPSCatalog
 from zope.component import getMultiAdapter
 from zope.component import getUtility
-from collective.elasticsearch.es import ElasticSearchCatalog
-from collective.elasticsearch.hook import index_batch
 
 
 logger = logging.getLogger("Plone")
@@ -158,11 +156,11 @@ def gen_audit_query(
     return query
 
 
-def index_in_es(obj):
-    catalog = api.portal.get_tool('portal_catalog')
-    es = ElasticSearchCatalog(catalog)
-    if es.enabled:
-        index_batch([], {IUUID(obj): obj}, [], es)
+#def index_in_es(obj):
+#    catalog = api.portal.get_tool('portal_catalog')
+#    es = ElasticSearchCatalog(catalog)
+#    if es.enabled:
+#        index_batch([], {IUUID(obj): obj}, [], es)
 
 
 def get_index_summary(index_name, terms, agg_agg_terms=None, filter_terms=None):